backend/requirements.txt
flask
flask-cors
google-cloud-vision
google-generativeai
python-dotenv
pillow